lpc: Add support for large read/write ops
On P9 the LPC is memory mapped. The LPC FW space used to map
the flash is generally acccessed to read or write rather large
amounts of data.

Its inefficient to go trough the whole dispatch and locking for
every single 4 bytes read or written.

This adds the ability to request reads or writes of larger
quantities for FW space. The implementation uses memcpy whose
current implementation in HostBoot will perform well for 8
bytes aligned accesses, but will downgrate to bytes
accesses otherwise.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>